# src/docs/swagger.yaml
basePath: /api/v1
definitions:
  booking.Availability:
    enum:
    - UNAVAILABLE
    - AVAILABLE
    - BOOKED
    type: string
    x-enum-varnames:
    - Unavailable
    - Available
    - Booked
  booking.AvailabilitySlot:
    properties:
      availability:
        $ref: '#/definitions/booking.Availability'
      location_id:
        type: string
      resource_type:
        type: string
      time_slot_id:
        type: string
    type: object
  booking.Booking:
    properties:
      confirmation_closed:
        type: string
      confirmation_open:
        type: string
      location_id:
        type: string
      resource_id:
        type: string
      show_confirm_button:
        type: boolean
      show_unbook_button:
        type: boolean
      time_slot:
        $ref: '#/definitions/booking.TimeSlot'
    type: object
...
  booking.ConfirmBookingRequest:
    properties:
      resourceId:
        type: string
    type: object
  booking.Resource:
    properties:
      availabilities:
        additionalProperties:
          additionalProperties:
            $ref: '#/definitions/booking.AvailabilitySlot'
          type: object
        type: object
      date:
        type: string
      id:
        type: string
      location_ids:
        items:
          type: string
        type: array
      name:
        type: string
      time_slots:
        items:
          $ref: '#/definitions/booking.TimeSlot'
        type: array
    type: object
  booking.TimeSlot:
    properties:
      from:
        type: string
      id:
        type: integer
      to:
        type: string
    type: object
  handlers.AvailabilityResponse:
    description: Response containing available time slots
    properties:
      availability:
        items:
          $ref: '#/definitions/booking.AvailabilitySlot'
        type: array
    type: object
  handlers.BookingsListResponse:
    description: Response containing list of bookings
    properties:
      bookings:
        items:
          $ref: '#/definitions/booking.Booking'
        type: array
    type: object
...
  handlers.ErrorResponse:
    description: Error response structure
    properties:
      error:
        example: Invalid request
        type: string
    type: object
...
  handlers.ResourcesListResponse:
    description: Response containing list of available resources
    properties:
      resources:
        items:
          $ref: '#/definitions/booking.Resource'
        type: array
    type: object
...
  handlers.SuccessResponse:
    description: Standard success response
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
host: localhost:5055
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A RESTful Web API for managing Kronox web resources, schedules, events,
    and bookings
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kronox API
  version: "1.0"
paths:
...
  /resources:
    get:
      consumes:
      - application/json
      description: Retrieve all available resources for booking across multiple school
        URLs
      parameters:
      - description: Bearer token (session ID)
        format: Bearer {session_id}
        in: header
        name: Authorization
        required: true
        type: string
      - description: School that request pertains to
        example: '"hkr"'
        in: query
        name: school
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of available resources
          schema:
            $ref: '#/definitions/handlers.ResourcesListResponse'
        "401":
          description: Session required
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all resources
      tags:
      - resources
  /resources/{resourceId}/availability:
    get:
      consumes:
      - application/json
      description: Get available time slots for a specific resource on a given date
        using path parameter
      parameters:
      - description: Bearer token (session ID)
        format: Bearer {session_id}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Resource ID to check availability for
        in: path
        name: resourceId
        required: true
        type: string
      - description: Date in YYYY-MM-DD format
        format: date
        in: query
        name: date
        required: true
        type: string
      - description: School that request pertains to
        example: '"hkr"'
        in: query
        name: school
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Available time slots for the resource
          schema:
            $ref: '#/definitions/handlers.AvailabilityResponse'
        "400":
          description: resourceId and date are required
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Session required
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get available resources (alternative endpoint)
      tags:
      - resources
...
  /resources/{resourceId}/bookings:
    get:
      consumes:
      - application/json
      description: Retrieve all active bookings for a specific resource
      parameters:
      - description: Bearer token (session ID)
        format: Bearer {session_id}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Resource ID to get bookings for
        in: path
        name: resourceId
        required: true
        type: string
      - description: School that request pertains to
        example: '"hkr"'
        in: query
        name: school
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of active bookings for the resource
          schema:
            $ref: '#/definitions/handlers.BookingsListResponse'
        "400":
          description: resourceId path parameter required
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Session required
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get active bookings for resource
      tags:
      - bookings
...
  /resources/bookings/{bookingId}/confirm:
    post:
      consumes:
      - application/json
      description: Confirm an existing resource booking with resource ID in request
        body
      parameters:
      - description: Bearer token (session ID)
        format: Bearer {session_id}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Booking ID to confirm
        in: path
        name: bookingId
        required: true
        type: string
      - description: Confirmation details
        in: body
        name: confirmation
        required: true
        schema:
          $ref: '#/definitions/booking.ConfirmBookingRequest'
      - description: School that request pertains to
        example: '"hkr"'
        in: query
        name: school
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Booking confirmed successfully
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Session required
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Confirm a booking
      tags:
      - bookings
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and session token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"